initramfs: Pre-patch with Magisk's want_initramfs hack
The majority of users use Magisk, including many developers. Our current
default case of *not* being patched for Magisk ("skip_initramfs") means
that users who *have* Magisk need to wait longer during the flashing
process for an expensive decompress + patch + recompress process.

We still need to use the skip_initramfs and want_initramfs strings to
be compatible with Magisk's uninstaller and prevent users from getting
stuck in recovery, so switch to a default case of "want_initramfs"
(patched for Magisk) instead. This will save the time spent on the
patching process for the majority of users. It also allows developers
who use Magisk to directly boot kernel images through fastboot without
needing to flash anything.

Note that this will require flasher adjustments to reverse the patching
logic and effectively act as a Magisk uninstaller for users who do not
have Magisk.

Signed-off-by: Danny Lin <danny@kdrag0n.dev>
Signed-off-by: Alex Winkowski <dereference23@outlook.com>
Signed-off-by: Forenche <prahul2003@gmail.com>
Signed-off-by: 7Soldier <reg.fm4@gmail.com>